+++ core/src/main/scala/org/broadinstitute/dsde/rawls/spendreporting/SpendReportingService.scala
@@ -306,6 +306,7 @@ class SpendReportingService(
     )
   }
 
+  // single-project report: query
   def getQuery(aggregations: Set[SpendReportingAggregationKeyWithSub], config: BillingProjectSpendExport): String = {
     // Unbox potentially many SpendReportingAggregationKeyWithSubs for query,
     // all of which have optional subAggregationKeys and convert to Set[SpendReportingAggregationKey]
@@ -332,9 +333,10 @@ class SpendReportingService(
     if (isBroadTable) spendReportingServiceConfig.defaultTimePartitionColumn else "_PARTITIONTIME"
   }
 
+  // consolidated report: query
   def getAllUserWorkspaceQuery(
     spendExportTable: String,
-    workspaces: Seq[GoogleProjectId],
+    billingProjectsByAccount: Map[RawlsBillingAccountName, Seq[GoogleProjectId]],
     pageSize: Int,
     offset: Int
   ): String = {
@@ -352,8 +354,8 @@ class SpendReportingService(
                        |  FROM
                        |    _BILLING_ACCOUNT_TABLE
                        |  where
-                       |    project.id in _PROJECT_ID_LIST AND
-                       |    _PARTITIONTIME BETWEEN @startDate AND @endDate
+                       |    _PARTITIONTIME BETWEEN @startDate AND @endDate and
+                       |    (_PROJECT_CLAUSE)
                        |  GROUP BY
                        |    project_id,
                        |    spend_category,
@@ -364,11 +366,7 @@ class SpendReportingService(
       baseQuery
         .replace("_PARTITIONTIME", timePartitionColumn)
         .replace("_BILLING_ACCOUNT_TABLE", spendExportTable)
-        .replace("_PROJECT_ID_LIST",
-                 workspaces
-                   .map(projectId => s""""${projectId}"""")
-                   .mkString("(", ", ", ")")
-        )
+        .replace("_PROJECT_CLAUSE", billingProjectsByAccountClause(billingProjectsByAccount))
 
     s"""WITH spend_categories AS (
        |$bpSubQuery
@@ -394,6 +392,17 @@ class SpendReportingService(
        |""".stripMargin.trim
   }
 
+  def billingProjectsByAccountClause(
+    billingProjectsByAccount: Map[RawlsBillingAccountName, Seq[GoogleProjectId]]
+  ): String =
+    billingProjectsByAccount
+      .map { case (billingAccount, projects) =>
+        val quotedProjects = projects.map(p => s"'${p.value}'")
+        s" (billing_account_id = '${billingAccount.withoutPrefix()}' and project.id in ${quotedProjects.mkString("(", ", ", ")")}) "
+      }
+      .mkString(" or ")
+
+  // single-project report: query parameters
   def setUpQuery(
     query: String,
     exportConf: BillingProjectSpendExport,
@@ -445,6 +454,7 @@ class SpendReportingService(
     bytesProcessedCounter += stats.getEstimatedBytesProcessed
   }
 
+  // shared between single-project and consolidated reports
   def getSpendReportableWorkspaceGoogleProjects(
     childContext: RawlsRequestContext
   ): Future[Map[RawlsBillingProjectName, Seq[Workspace]]] =
@@ -462,6 +472,7 @@ class SpendReportingService(
         workspaceServiceConstructor(childContext).getGCPWorkspacesByBillingProjects(validWorkspaceIds)
       }
 
+  // single-project report: step 3
   def getSpendForGCPBillingProject(
     project: RawlsBillingProjectName,
     start: DateTime,
@@ -480,6 +491,13 @@ class SpendReportingService(
           .toMap
       }
 
+      _ = if (projectNames.isEmpty) {
+        throw RawlsExceptionWithErrorReport(
+          StatusCodes.NotFound,
+          s"no spend data found for billing project ${project.value} between dates ${toISODateString(start)} and ${toISODateString(end)}"
+        )
+      }
+
       query = getQuery(aggregations, spendExportConf)
       queryJob = setUpQuery(query, spendExportConf, start, end, projectNames)
 
@@ -496,6 +514,7 @@ class SpendReportingService(
     }
   }
 
+  // single-project report: entry point
   def getSpendForBillingProject(
     project: RawlsBillingProjectName,
     start: DateTime,
@@ -508,6 +527,7 @@ class SpendReportingService(
       report <- getReportData(billingProject.get, project, start, end, aggregations)
     } yield report
 
+  // single-project report: step 2
   private def getReportData(billingProject: RawlsBillingProject,
                             project: RawlsBillingProjectName,
                             start: DateTime,
@@ -552,6 +572,7 @@ class SpendReportingService(
           Future.failed(RawlsExceptionWithErrorReport(ErrorReport(StatusCodes.InternalServerError, ex)))
       }
 
+  // consolidated report: entry point
   def getSpendForAllWorkspaces(
     start: DateTime,
     end: DateTime,
@@ -561,73 +582,54 @@ class SpendReportingService(
     traceFutureWithParent("getSpendForAllWorkspaces", ctx) { childContext =>
       validateReportParameters(start, end)
       for {
+        // get spend-reportable workspaces, grouped by their spend export config
+        // Map[BillingProjectSpendExport, Seq[Workspace]]
         billingMap <- getBillingWithSpendPermission(childContext)
         _ = if (billingMap.isEmpty) {
-          return Future.successful(None)
+          throw new RawlsExceptionWithErrorReport(
+            ErrorReport(StatusCodes.NotFound, "No workspaces eligible for spend report")
+          )
         }
-        projectNames: Map[GoogleProjectId, WorkspaceName] = billingMap.values.flatten.toMap
-        projectIds = projectNames.keySet.map(_.value)
-        tableToProjectIdsMap: Map[String, Seq[GoogleProjectId]] = billingMap.map { case (table, workspaces) =>
-          table -> workspaces.map(_._1)
-        }
-        startLocalDateTime =
-          LocalDateTime.ofInstant(java.time.Instant.ofEpochMilli(start.getMillis), ZoneId.systemDefault())
-        endLocalDateTime =
-          LocalDateTime.ofInstant(java.time.Instant.ofEpochMilli(end.getMillis), ZoneId.systemDefault())
-        cacheResults <- workspaceSpendReportRepository.getWorkspaceSpendReports(projectIds,
-                                                                                startLocalDateTime,
-                                                                                endLocalDateTime
-        )
-        combinedResults <-
-          if (cacheResults.size == projectIds.size) {
-            val hasDataAvailable = cacheResults.filter(cached => cached.isDataAvailable)
-            Future.successful(Some(WorkspaceSpendReportRecord.toSpendReportingResults(hasDataAvailable, projectNames)))
-          } else {
-            getUncachedSpendForAllWorkspaces(tableToProjectIdsMap,
-                                             projectNames,
-                                             start,
-                                             end,
-                                             pageSize,
-                                             offset,
-                                             childContext
-            )
-          }
-      } yield combinedResults
-    }
 
-  def getUncachedSpendForAllWorkspaces(tableToProjectIdsMap: Map[String, Seq[GoogleProjectId]],
-                                       projectNames: Map[GoogleProjectId, WorkspaceName],
-                                       start: DateTime,
-                                       end: DateTime,
-                                       pageSize: Int,
-                                       offset: Int,
-                                       childContext: RawlsRequestContext
-  ): Future[Option[SpendReportingResults]] =
-    for {
-      results <- Future.sequence(tableToProjectIdsMap.map { case (spendExportTable, projects) =>
-        val query = getAllUserWorkspaceQuery(spendExportTable, projects, pageSize, offset)
-        val queryJob = setUpAllUserWorkspaceQuery(query, start, end)
-        runBigQueryJob(queryJob, childContext)
-          .map { result =>
-            result.getValues.asScala.toList match {
-              case Nil => None
-              case rows =>
-                val crossBillingResults =
-                  extractCrossBillingProjectSpendReportingResults(rows, start, end, projectNames)
-                workspaceSpendReportRepository.insertSpendReportResults(crossBillingResults)
-                Some(crossBillingResults)
+        // find the distinct export table names in our billingMap
+        distinctTableNames: Set[Option[String]] = billingMap.keys.map(_.spendExportTable).toSet
+
+        // for each distinct export table name, generate a query.
+        results <- Future.sequence(distinctTableNames.map { spendExportTable =>
+          // find the subset of the billingMap that we'll use in this query
+          val billingMapForQuery = billingMap.filter(_._1.spendExportTable == spendExportTable)
+          // use default export table if none is specified
+          val tableNameForQuery = spendExportTable.getOrElse(spendReportingServiceConfig.defaultTableName)
+          // find the (billingAccount, billingProjects) pairs for this query
+          val billingProjectsByAccount: Map[RawlsBillingAccountName, Seq[GoogleProjectId]] =
+            billingMapForQuery
+              .groupMap(_._1.billingAccountId)(_._2)
+              .view
+              .mapValues(_.flatten.map(_.googleProjectId).toSeq)
+              .toMap
+
+          // and finally, generate a map of GoogleProjectId->WorkspaceName for extracting the spend results
+          val workspaceNamesByProjectId: Map[GoogleProjectId, WorkspaceName] = billingMapForQuery.values.flatten
+            .map(ws => ws.googleProjectId -> ws.toWorkspaceName)
+            .toMap
+
+          val query = getAllUserWorkspaceQuery(tableNameForQuery, billingProjectsByAccount, pageSize, offset)
+          val queryJob = setUpAllUserWorkspaceQuery(query, start, end)
+          runBigQueryJob(queryJob, childContext)
+            .map { result =>
+              result.getValues.asScala.toList match {
+                case Nil => None
+                case rows =>
+                  Some(extractCrossBillingProjectSpendReportingResults(rows, start, end, workspaceNamesByProjectId))
+              }
             }
-          }
-          .recoverWith { case ex: Throwable =>
-            logger.warn(s"Error fetching results from BigQuery: ${ex.getMessage}")
-            Future.successful(None)
-          }
-      })
-      _ = if (results.nonEmpty) {
-        insertRecordsWithMissingSpendData(results, projectNames, start, end)
-      }
-      combinedResults = results.flatten.reduceOption((acc, res) => acc + res)
-    } yield combinedResults
+            .recoverWith { case ex: Throwable =>
+              logger.warn(s"Error fetching results from BigQuery: ${ex.getMessage}")
+              Future.successful(None)
+            }
+        })
+        combinedResults = results.flatten.reduceOption((acc, res) => acc + res)
+      } yield combinedResults
 
   def insertRecordsWithMissingSpendData(results: Iterable[Option[SpendReportingResults]],
                                         projectNames: Map[GoogleProjectId, WorkspaceName],
@@ -674,28 +676,30 @@ class SpendReportingService(
 
   def getBillingWithSpendPermission(
     parentContext: RawlsRequestContext
-  ): Future[Map[String, Seq[(GoogleProjectId, WorkspaceName)]]] =
+  ): Future[Map[BillingProjectSpendExport, Seq[Workspace]]] =
     traceFutureWithParent("getBillingWithSpendPermission", parentContext) { childContext =>
       for {
-        groupedWorkspaces <- getSpendReportableWorkspaceGoogleProjects(childContext)
-        // Only use the BPs we know exist in the DB and are GCP
+        // Retrieve the workspaces for which the user is allowed to report on spend;
+        // these are grouped by billing project
+        workspacesByBillingProjectName <- getSpendReportableWorkspaceGoogleProjects(childContext)
+        // Retrieve the spend-export configs for each of these billing projects;
+        // this verifies the billing projects exist in the DB and are GCP
         spendConfigs <-
-          if (groupedWorkspaces.isEmpty) {
+          if (workspacesByBillingProjectName.isEmpty) {
             Future.successful(Seq.empty[BillingProjectSpendExport])
           } else {
-            getSpendExportConfigurations(groupedWorkspaces.keys.toList)
+            getSpendExportConfigurations(workspacesByBillingProjectName.keys.toList)
           }
-        groupedByTable = spendConfigs.groupBy(
-          _.spendExportTable.getOrElse(spendReportingServiceConfig.defaultTableName)
-        )
-        combinedResults = groupedByTable.map { case (table, configs) =>
-          val combinedWorkspaces = configs.flatMap { config =>
-            groupedWorkspaces
-              .getOrElse(RawlsBillingProjectName(config.billingProjectName.value), Seq.empty)
-              .map(ws => (ws.googleProjectId, ws.toWorkspaceName))
+
+        // Match the workspaces to their spend configs. Filter out any spend configs which
+        // have no workspaces (this is not expected to happen)
+        workspacesBySpendConfig = spendConfigs
+          .map { config =>
+            config -> workspacesByBillingProjectName.getOrElse(config.billingProjectName, Seq.empty[Workspace])
           }
-          table -> combinedWorkspaces
-        }
-      } yield combinedResults
+          .toMap
+          .filter(_._2.nonEmpty)
+
+      } yield workspacesBySpendConfig
     }
 }