+++ core/src/test/scala/org/broadinstitute/dsde/rawls/webservice/WorkspaceApiListOptionsSpec.scala
@@ -75,7 +75,8 @@ class WorkspaceApiListOptionsSpec extends ApiServiceSpec {
       )
     )
 
-    val billingProject = RawlsBillingProject(RawlsBillingProjectName("ns"), CreationStatuses.Ready, None, None)
+    val billingProject =
+      RawlsBillingProject(UUID.randomUUID(), RawlsBillingProjectName("ns"), CreationStatuses.Ready, None, None)
 
     val workspaceName = WorkspaceName(billingProject.projectName.value, "testworkspace")